Chhaya_new_backend/requirement.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
distlib==0.3.8
Django==5.1.3
django-cors-headers==4.3.1
django-filter==24.3
django-leaflet==0.31.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.0
fiona 
GDAL 
geopandas==0.14.4
git-filter-repo==2.45.0
idna==3.7
inflection==0.5.1
leaflet==0.0.3
lxml==5.3.0
MouseInfo==0.1.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.3.2
psycopg2==2.9.9
PyAutoGUI==0.9.54
pycryptodome==3.20.0
pyfiglet==1.0.2
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.9.0
pyproj==3.6.1
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-redis-cache==4.0.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.2.0
reportlab==4.2.2
requests==2.32.3
shapely==2.0.4
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
termcolor==2.4.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.4
virtualenvwrapper-win==1.2.7
XlsxWriter==3.2.0
